refactor(#16): 컴포넌트 단일 책임 적용
- ProtectedRoute를 App.tsx 인라인에서 독립 컴포넌트로 분리
- NicknameForm 컴포넌트 완성 (UI + useGuestSession 훅 연결)
- Home.tsx에서 비즈니스 로직 제거, NicknameForm 배치만 담당
- App.tsx에서 불필요한 import React 제거
